SSHT.jl: Fast and exact spin spherical harmonic transforms

This package is a Julia wrapper for the SSHT library. It calculates spin spherical harmonic transforms.

Most functions come in two versions, one that mutates its arguments and one that allocates its output.

Spin-weighted Spherical Harmonic transforms

The functions with dh in their names use the Driscoll & Healy quadrature points. With L = lmax+1 modes there are nphi * ntheta quadrature points on the sphere, with nphi = 2L-1 and ntheta = 2L. The points are equispaced in the angles theta (latitude) and phi (longitude), and they straddle (avoid) the poles.

SSHT.core_dh_inverse_sov! — Function

SSHT.core_dh_inverse_sov!(f::AbstractArray{Complex{Float64},2},
                          flm::AbstractVector{Complex{Float64}},
                          L::Integer,
                          spin::Integer,
                          verbosity::Integer=0)

Evaluate spin-weighted spherical harmonic coefficients flm with spin-weight spin. L = lmax+1 is the number of modes in flm.

The arrays flm and f must have the following sizes:

nphi = SSHT.sampling_dh_nphi(L)
ntheta = SSHT.sampling_dh_ntheta(L)
length(flm) == L^2
size(f) == (nphi, ntheta)

See the SSHT reference for details.

See also: core_dh_inverse_sov!, core_dh_inverse_sov_real!, core_dh_forward_sov!.

SSHT.core_dh_inverse_sov — Function

f = SSHT.core_dh_inverse_sov(flm::AbstractVector{Complex{Float64}},
                             L::Integer,
                             spin::Integer,
                             verbosity::Integer=0)
f::Array{Complex{Float64},2}

Evaluate spin-weighted spherical harmonic coefficients flm with spin-weight spin. L = lmax+1 is the number of modes in flm.

The array flm must have the length L^2. The result f will have the following size:

nphi = SSHT.sampling_dh_nphi(L)
ntheta = SSHT.sampling_dh_ntheta(L)
size(f) == (nphi, ntheta)

See the SSHT reference for details.

See also: core_dh_inverse_sov, core_dh_inverse_sov_real, core_dh_forward_sov.

Missing docstring.

Missing docstring for SSHT.core_dh_inverse_sov_real!. Check Documenter's build log for details.

Missing docstring.

Missing docstring for SSHT.core_dh_inverse_sov_real. Check Documenter's build log for details.

Missing docstring.

Missing docstring for SSHT.core_dh_forward_sov!. Check Documenter's build log for details.

Missing docstring.

Missing docstring for SSHT.core_dh_forward_sov. Check Documenter's build log for details.

Missing docstring.

Missing docstring for SSHT.core_dh_forward_sov_real!. Check Documenter's build log for details.

Missing docstring.

Missing docstring for SSHT.core_dh_forward_sov_real. Check Documenter's build log for details.

Derivatives ð (eth) and ð̄ (eth-bar)

See Wikipedia for the definition of these operators.

Missing docstring.

Missing docstring for SSHT.eth!. Check Documenter's build log for details.

Missing docstring.

Missing docstring for SSHT.eth. Check Documenter's build log for details.

Missing docstring.

Missing docstring for SSHT.ethbar!. Check Documenter's build log for details.

Missing docstring.

Missing docstring for SSHT.ethbar. Check Documenter's build log for details.

Helper functions

Missing docstring.

Missing docstring for SSHT.sampling_dh_n. Check Documenter's build log for details.

Missing docstring.

Missing docstring for SSHT.sampling_dh_nphi. Check Documenter's build log for details.

Missing docstring.

Missing docstring for SSHT.sampling_dh_ntheta. Check Documenter's build log for details.

Missing docstring.

Missing docstring for SSHT.sampling_dh_p2phi. Check Documenter's build log for details.

Missing docstring.

Missing docstring for SSHT.sampling_dh_t2theta. Check Documenter's build log for details.

Missing docstring.

Missing docstring for SSHT.sampling_weight_dh. Check Documenter's build log for details.

Missing docstring.

Missing docstring for SSHT.sampling_elm2ind. Check Documenter's build log for details.

Missing docstring.

Missing docstring for SSHT.sampling_ind2elm. Check Documenter's build log for details.